backend/requirements.txt
fastapi
uvicorn
python-telegram-bot[httpx]
colorlog
pytz
pydantic
